4.17 - 4.21

+++ osa4/blogilista/controllers/blogs.js
@@ -1,9 +1,15 @@
 const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
+const User = require("../models/user");
+const jwt = require("jsonwebtoken");
+const config = require("../utils/config");
 
 blogsRouter.get("/", async (request, response, next) => {
   try {
-    const blogs = await Blog.find({});
+    const blogs = await Blog.find({}).populate("user", {
+      username: 1,
+      name: 1
+    });
     response.json(blogs);
   } catch (error) {
     next(error);
@@ -11,11 +17,28 @@ blogsRouter.get("/", async (request, response, next) => {
 });
 
 blogsRouter.post("/", async (request, response, next) => {
-  const blog = new Blog(request.body);
+  const body = request.body;
 
   try {
-    const result = await blog.save();
-    response.status(201).json(result);
+    const decodedToken = jwt.verify(request.token, config.JWT_SECRET);
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: "Token missing or invalid" });
+    }
+
+    const user = await User.findById(decodedToken.id);
+
+    const blog = new Blog({
+      title: body.title,
+      author: body.author,
+      url: body.url,
+      likes: body.likes,
+      user: user._id
+    });
+
+    const savedBlog = await blog.save();
+    user.blogs = user.blogs.concat(savedBlog._id);
+    await user.save();
+    response.status(201).json(savedBlog);
   } catch (error) {
     next(error);
   }
@@ -41,6 +64,18 @@ blogsRouter.put("/:id", async (request, response, next) => {
 
 blogsRouter.delete("/:id", async (request, response, next) => {
   try {
+    const decodedToken = jwt.verify(request.token, config.JWT_SECRET);
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: "Token missing or invalid" });
+    }
+    const user = await User.findById(decodedToken.id);
+    const blog = await Blog.findById(request.params.id);
+    if (user._id.toString() !== blog.user.toString()) {
+      return response
+        .status(403)
+        .json({ error: "Not authorized to remove other user's blog" });
+    }
+
     await Blog.findByIdAndDelete(request.params.id);
     response.status(204).end();
   } catch (error) {
